move jobs list to a hook
